Move: Added Folder Name Checking Mechanism

+++ MoveV1_XP.bat
@@ -8,30 +8,41 @@ setlocal enabledelayedexpansion
 set "sourceDir=F:"
 set "destinationDir=\\vmware-host\Shared Folders\tempshared"
 
-:: Get the date from the folders' name
+:: Get the date from the folders' name, ONLY if it is in ddMMMyy format, else skip
 for /d %%d in (%sourceDir%\*) do (
     set folderName=%%~nxd
-    set year=20!folderName:~5,2!
-    set month=!folderName:~2,3!
-    set day=!folderName:~0,2!
-    echo Year:!year! Month:!month! Day:!day!
-    
-    REM Create the folder with today's date in the destination directory
-    set "datePath=%destinationDir%\!year!\!month!\!day!"
 
-    if not exist "!datePath!" (
-        mkdir "!datePath!"
-    )
+    REM Is the folder name null?
+    if "!folderName!" neq "" (
+        set year=20!folderName:~5,2!
+        set month=!folderName:~2,3!
+        set day=!folderName:~0,2!
 
-    REM Move files from source to newly created folder in destination
-    xcopy "%%d\*" "!datePath!" /e /d /q /y
+        REM Check if the date is valid
+        if "!day!" GEQ "01" if "!day!" LEQ "31" (
+            echo Processing folder: %%d
+            echo Year:!year! Month:!month! Day:!day!
+            
+            REM Create the date folder in the destination directory
+            set "datePath=%destinationDir%\!year!\!month!\!day!"
 
-    REM Remove the source folder
-    if !errorlevel! equ 0 (
-        echo Copy success, deleting folder: %%d
-        rmdir "%%d" /s /q
-    ) else (
-        echo Copy failed, not deleting folder: %%d
+            if not exist "!datePath!" (
+                mkdir "!datePath!"
+            )
+
+            REM Move files from source to newly created folder in destination
+            xcopy "%%d\*" "!datePath!" /e /d /q /y
+
+            REM Remove the source folder
+            if !errorlevel! equ 0 (
+                echo Copy success, deleting folder: %%d
+                rmdir "%%d" /s /q
+            ) else (
+                echo Copy failed, not deleting folder: %%d
+            )
+        ) else (
+            echo Skipping invalid folder: %%d
+        )
     )
 )
 endlocal
